seed/requirements.txt
supabase==2.11.0
python-dotenv==1.0.1
